requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
qdrant-client
neo4j
httpx
hello-agents==0.2.8
dashscope

sentence-transformers
